fix: Set up the backend to run the project
Upgraded few packages, added necessary code to run the backend properly. Set up .gitignore to prevent the build files and passwords being uploaded into the repo. Set up CORS & made to allow '*' for origins. Implemented GET /categories endpoint to check if the connection to the database is working with the API call.